requirements.txt
fastapi==0.115.12 
python-dotenv
psycopg2
pandas